todostableエリアをcomponent分割
