+++ app/components/Todostable.module.scss
@@ -1,0 +1,65 @@
+@use "../variables.module.scss" as vars;
+
+.tableSetting {
+  margin-bottom: 32px;
+  display: flex;
+  justify-content: flex-end;
+  padding: 16px;
+}
+.tableHeader {
+  display: flex;
+  border-bottom: 1px solid vars.$borderColor;
+  .tableHeaderCell {
+    padding: 8px 16px;
+    font-size: 12px;
+    flex: 1 1;
+    border-right: 1px solid vars.$borderColor;
+  }
+  .tableHeaderCellTask {
+    padding: 8px 16px;
+    font-size: 12px;
+    flex: 4 1;
+    border-right: 1px solid vars.$borderColor;
+  }
+}
+.tableRow {
+  display: flex;
+  justify-content: space-between;
+  border-bottom: 1px solid vars.$borderColor;
+  transition: 1s;
+}
+.tableCell {
+  padding: 16px;
+  flex: 1 1;
+  border-right: 1px solid vars.$borderColor;
+  display: flex;
+  align-items: center;
+  .checkbox {
+    cursor: pointer;
+    border: 1px solid vars.$fontWeakText;
+    border-radius: 50%;
+    height: 28px;
+    width: 28px;
+    padding: 4px;
+    color: vars.$fontColorLight;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  .checkboxChecked {
+    @extend .checkbox;
+    background-color: vars.$primartColor;
+  }
+}
+.tableCellCenter {
+  @extend .tableCell;
+  justify-content: center;
+  .trashIcon {
+    cursor: pointer;
+    font-size: 1rem;
+  }
+}
+.tableCellTask {
+  @extend .tableCell;
+  flex: 4 1;
+}
